refactor(tests): extract driver integration tests into a separate file
Move the five integration tests from the monolithic `driver.rs` into a dedicated `driver_part_02_tests.rs` module, keeping only the `include!` directive in the original file. This reduces the main test file by 280 lines and improves maintainability by grouping related test scenarios in their own module.

Auto-committed-on: dragonfly
